requirements.txt
Django==1.9.5
django-allauth==0.25.2
django-suit==0.2.18
django-tinymce==2.3.0
oauthlib==1.1.2
Pillow==3.2.0
PyJWT==1.4.0
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
